Tags, virtualenvs, env var, center selection
All mixed together as I forget to commit in my rambling explorations of
haskell.

This introduces an optional "tags" entry into the locations listed in
the yaml file. This is for extra meta-data which we use to highlight the
Python virtual env for each location if there.

We then change the output to check for the virtualenv and write out the
activate line to the output file, also check for a current virtualenv and
write out a decactivate line, as well as the "cd" command which is what
jump is meant to do.

Could be a lot nicer I imagine but that will come with time.

We also schedule a change to center the selection on the list at start
up. This cannot be done (from trial and error) before the UI loop as
kicked in so we use "schudule" to add it to the loop to be run once
everything is up and running.

And fixed various ghc compiler warnings

- No type signatures
- Import semantics (Extra ()s)
- Ignoring statement result in do section (_ <- ...)
- Unused record names (getDirectory getTags)
- Variable shadowing (border)
- Missing patterns (Nothing)